+++ api/client/client.go
@@ -337,7 +337,7 @@ func (bc *baseClient) Stop(ctx context.Context) error {
 	return err
 }
 
-func (bc *baseClient) WidthMillis(ctx context.Context) (int, error) {
+func (bc *baseClient) WidthMillis(ctx context.Context) (float64, error) {
 	debug.PrintStack()
 	return 0, errUnimplemented
 }
